Grounded_Segment_Anything/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
transformers>=4.20.0
timm>=0.6.0
pycocotools>=2.0.0
tqdm>=4.60.0
scipy>=1.7.0
scikit-image>=0.18.0 